requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
PyMySQL
Flask-Session